+++ CHANGELOG.md
@@ -19,6 +19,7 @@ For details about compatibility between different releases, see the **Commitment
 - Experimental support for BasicStation GPS timestamps which use the wrong precision (milliseconds instead of microseconds).
   - The Gateway Server will attempt to determine the correct GPS timestamp from the provided `gpstime` based on the time at which the upstream message has been received.
   - This workaround will be available until the related gateway vendors will release patches for this issue.
+- Firmware version of The Things Kickstarter Gateway are written to the gateway attributes upon receiving a valid status message.
 
 ### Changed